Detach process when running tests
